fix get rank

+++ BE/src/main/java/com/arkdev/z9tkvtu/dto/Response/UserRankResponse.java
@@ -1,10 +1,9 @@
 package com.arkdev.z9tkvtu.dto.Response;
 
 import java.io.Serializable;
-import java.util.UUID;
 
 public record UserRankResponse(
         String username,
         Integer totalScore,
-        Integer totalTime) implements Serializable {
+        long totalTime) implements Serializable {
 }
